Move modern craft and forge into crafting module
