+++ lib/crafting.cfg
@@ -27,3 +27,16 @@ crafting.level.gate.multiplier=3
 # required ingredient. Must be greater than zero.
 # Historical default: 1.0 (no change).
 crafting.material.quantity.multiplier=1.0
+
+# Opt-in/rollback gates for the modern commands. `0` leaves legacy smith-NPC
+# forging untouched; it disables only `craft` or the recipe-based `forge`.
+crafting.craft.enabled=1
+crafting.forge.enabled=1
+
+# Existing support-item vnums. Positive values only are accepted. These are
+# content references, not engine storage invariants. Craft and Forge may use
+# distinct vnums even when current content maps their essences together.
+crafting.craft.essence.vnum=400211
+crafting.craft.tool.vnum=400224
+crafting.forge.essence.vnum=400211
+crafting.forge.tool.vnum=400223
