+++ src/actnew.c
@@ -3616,525 +3616,9 @@ void do_vote(P_char ch, char *arg, int cmd)
 
 void do_craft(P_char ch, char *argument, int cmd)
 {
-	char  buf1[MAX_STRING_LENGTH];
-	char  first[MAX_INPUT_LENGTH];
-	char  second[MAX_INPUT_LENGTH];
-	char  rest[MAX_INPUT_LENGTH];
-	int   i      = 0;
-	int   choice = 0;
-	P_obj hammer, foundry;
-
-	/***DISPLAYRECIPES STUFF***/
-	char          tempdesc[MAX_INPUT_LENGTH];
-	char          short_desc[MAX_STRING_LENGTH];
-	char          keywords[MAX_INPUT_LENGTH];
-	char          buffer[256];
-	long          choice2;
-	int           selected = 0;
-	int          *recipes;
-	int           recipe_count;
-	P_obj         tobj;
-
-	/* Dunno the difference between this and above?
-	   P_obj    craft_obj1, craft_obj2, craft_obj3, obj;
-	   P_obj    t_obj, nextobj;
-	   int      i, bits, j, in_room, material_type, item_type, howmany,
-	   weapon_types, slot;
-	   bool     equipped;
-	   P_char   victim = NULL;
-	   char     Gbuf1[MAX_STRING_LENGTH];
-	   char     Gbuf2[MAX_STRING_LENGTH];
-	   char    *r_str;
-
-	   equipped = FALSE;
-	   obj = 0;
-	   */
-
-	if (!GET_CHAR_SKILL(ch, SKILL_CRAFT))
-	{
-		act("You do not know how to &+rcraft&n items.", FALSE, ch, 0, 0, TO_CHAR);
-		return;
-	}
-
-	recipes = crafting_get_player_recipes(ch, &recipe_count);
-	if (recipes == NULL || recipe_count == 0)
-	{
-		send_to_char("You dont know any recipes yet.\r\n", ch);
-		free(recipes);
-		return;
-	}
-
-	// If no arguments, display syntax and list of known recipes.
-	if (!argument || !*argument)
-	{
-		send_to_char("&+wcraft Syntax:\n&n", ch);
-		send_to_char("&+w(craft info <number> - list required materials to craft the item.)\n&n", ch);
-		send_to_char("&+w(craft stat <number> - display properties of the item.)\n&n", ch);
-		send_to_char("&+w(craft make <number> - create the item.)\n&n", ch);
-		send_to_char("&+yYou know the following recipes:\n&n", ch);
-		send_to_char("----------------------------------------------------------------------------\n", ch);
-		send_to_char("&+BRecipe Number		              &+MItem&n\n\r", ch);
-		// Walk through each recipe and display its number and short description.
-		for (i = 0; i < recipe_count; i++)
-		{
-			tobj = read_object(recipes[i], VIRTUAL);
-			if (tobj == NULL)
-			{
-				logit(LOG_DEBUG, "do_craft: '%s' has bad recipe vnum %d.", J_NAME(ch), recipes[i]);
-				continue;
-			}
-			snprintf(buffer, 256, "   &+W%-22d&n%s&n\n", recipes[i], tobj->short_description);
-			page_string(ch->desc, buffer, 1);
-			send_to_char("----------------------------------------------------------------------------\n", ch);
-			extract_obj(tobj);
-		}
-		free(recipes);
-		return;
-	}
-
-	// If no argument, no sense in chopping it up, so moved this past the !arg check.
-	half_chop(argument, first, rest);
-	half_chop(rest, second, rest);
-	choice2 = atoi(second);
-
-	// Walk through the list of recipes and look for choice2.
-	if (choice2)
-	{
-		for (i = 0; i < recipe_count; i++)
-		{
-			if (recipes[i] == choice2)
-			{
-				selected = recipes[i];
-				break;
-			}
-		}
-	}
-	free(recipes);
-
-	// If they picked a recipe that wasn't in their book...
-	if (choice2 != 0 && selected == 0)
-	{
-		send_to_char("You dont appear to have that &+Wrecipe&n in your list.&n\n", ch);
-		return;
-	}
-
-	if (is_abbrev(first, "stat"))
-	{
-		if (choice2 == 0)
-		{
-			send_to_char("What &+Wrecipe&n would you like &+ystatistics&n about?\n", ch);
-			return;
-		}
-		tobj = read_object(selected, VIRTUAL);
-		send_to_char("&+yYou open your &+Ltome &+yof &+Ycra&+yftsm&+Lanship &+yand examine the &+Litem&n.\n", ch);
-		spell_identify(GET_LEVEL(ch), ch, 0, 0, 0, tobj);
-		extract_obj(tobj);
-		return;
-	}
-	else if (is_abbrev(first, "info"))
-	{
-		if (choice2 == 0)
-		{
-			send_to_char("What &+Wrecipe&n would you like &+yinformation&n about?\n", ch);
-			return;
-		}
-
-		tobj = read_object(selected, VIRTUAL);
-		if (!tobj)
-		{
-			debug("Couldn't load object vnum %d.", selected);
-			send_to_char("Couldn't create the object. :(\n\r", ch);
-			return;
-		}
-
-		struct crafting_plan plan;
-		if (!crafting_build_plan(tobj, &plan))
-		{
-			send_to_char("Could not figure out what this is made out of !?  Can bug it if you want.\n\r", ch);
-			debug("Couldn't build crafting plan for object: '%s' %d.", tobj->short_description, selected);
-			extract_obj(tobj);
-			return;
-		}
-		int iVal                   = plan.item_value;
-		int lowQualityMaterialVnum = plan.low_material_vnum;
-
-		if (lowQualityMaterialVnum <= 0)
-		{
-			send_to_char("Could not figure out what this is made out of !?  Can bug it if you want.\n\r", ch);
-			debug("Couldn't get start material for object: '%s' %d.", tobj->short_description, selected);
-			extract_obj(tobj);
-			return;
-		}
-
-		int numHighest = plan.high_material_count;
-		int numLowest = plan.low_material_count;
-
-		P_obj matLowest, matHighest;
-
-		matLowest  = read_object(lowQualityMaterialVnum, VIRTUAL);
-		matHighest = read_object(lowQualityMaterialVnum + 4, VIRTUAL);
-
-		if (matLowest == NULL || matHighest == NULL)
-		{
-			send_to_char("Could not figure out what this is made out of !?  Can bug it if you want.\n\r", ch);
-			debug("Couldn't load a material: matLowest(%s) or matHighest(%s) for object '%s' %d.",
-			      (matLowest == NULL) ? "NULL" : matLowest->short_description,
-			      (matHighest == NULL) ? "NULL" : matHighest->short_description,
-			      tobj->short_description,
-			      selected);
-			extract_obj(tobj);
-			return;
-		}
-
-		if (numLowest == 0)
-		{
-			send_to_char("&+yYou open your &+Ltome &+yof &+Ycra&+yftsm&+Lanship &+yand examine the &+Litem&n.\n", ch);
-			snprintf(buf1, MAX_STRING_LENGTH, "To craft this item, you will need %d of %s.\r\n&n", numHighest, matHighest->short_description);
-			page_string(ch->desc, buf1, 1);
-		}
-		else
-		{
-			send_to_char("&+yYou open your &+Ltome &+yof &+Ycra&+yftsm&+Lanship &+yand examine the &+Litem&n.\n", ch);
-			snprintf(buf1, MAX_STRING_LENGTH, "To craft this item, you will need %d of %s and %d of %s.\r\n&n", numHighest, matHighest->short_description, numLowest, matLowest->short_description);
-			page_string(ch->desc, buf1, 1);
-		}
-
-		/* It will never happen that we don't need at least one of matHighest since min itemvalue is 1,
-		 *   then we add 4 (total 5) and divide by 5 (result minimum 1).
-		 * But, in case we change that, I'm leaving the code here. 2/10/2015
-		      send_to_char("&+yYou open your &+Ltome &+yof &+Ycra&+yftsm&+Lanship &+yand examine the &+Litem&n.\n", ch);
-		      snprintf(buf1, MAX_STRING_LENGTH, "To craft this item, you will need %d of %s.\r\n&n", numLowest, matLowest->short_description);
-		      page_string(ch->desc, buf1, 1);
-		*/
-
-		if (has_affect(tobj))
-		{
-			send_to_char("...as well as &+W1 &nof &+ma &+Mm&+Ya&+Mg&+Yi&+Mc&+Ya&+Ml &+messence&n due to the &+mmagical &nproperties this item possesses.\r\n", ch);
-		}
-
-		// It's safe to assume tobj exists since we checked after the read_object call.
-		extract_obj(tobj);
-		// The same follows for matLowest and matHighest.
-		extract_obj(matLowest);
-		extract_obj(matHighest);
-		return;
-	}
-	else if (is_abbrev(first, "make"))
-	{
-		if (choice2 == 0)
-		{
-			send_to_char("What &+Witem&n are you attempting to craft?\n", ch);
-			return;
-		}
-
-		tobj = read_object(selected, VIRTUAL);
-		if (!tobj)
-		{
-			debug("Couldn't load object vnum %d.", selected);
-			send_to_char("Couldn't create the object. :(\n\r", ch);
-			return;
-		}
-
-		struct crafting_plan plan;
-		if (!crafting_build_plan(tobj, &plan))
-		{
-			send_to_char("Could not figure out what this is made out of !?  Can bug it if you want.\n\r", ch);
-			debug("Couldn't build crafting plan for object: '%s' %d.", tobj->short_description, selected);
-			extract_obj(tobj);
-			return;
-		}
-		int iVal = plan.item_value;
-		if (iVal > GET_LEVEL(ch) * crafting_level_gate_multiplier() || IS_OBJ_STAT2(tobj, ITEM2_QUESTITEM))
-		{
-			act("You look at the recipe for $p&n, but can't seem to discern how to make it.  &+mHow strange.&N", FALSE, ch, tobj, 0, TO_CHAR);
-			extract_obj(tobj);
-			return;
-		}
-
-		int lowQualityMaterialVnum  = plan.low_material_vnum;
-		int highQualityMaterialVnum = plan.high_material_vnum;
-		int numHighest = plan.high_material_count;
-		int numLowest = plan.low_material_count;
-
-		P_obj matLowest, matHighest;
-
-		matLowest  = read_object(lowQualityMaterialVnum, VIRTUAL);
-		matHighest = read_object(highQualityMaterialVnum, VIRTUAL);
-
-		if (matLowest == NULL || matHighest == NULL)
-		{
-			send_to_char("Could not figure out what this is made out of !?  Can bug it if you want.\n\r", ch);
-			debug("Couldn't load a material: matLowest(%s) or matHighest(%s) for object '%s' %d.",
-			      (matLowest == NULL) ? "NULL" : matLowest->short_description,
-			      (matHighest == NULL) ? "NULL" : matHighest->short_description,
-			      tobj->short_description,
-			      selected);
-			extract_obj(tobj);
-			return;
-		}
-
-		// If we're going to require multiple essences, we need to change this to int numAffects ...
-		bool hasAffect = has_affect(tobj);
-
-		int invLowMats  = 0; // Number of lowest quality materials in inventory.
-		int invHighMats = 0; // Number of highest quality materials in inventory.
-		int invEssences = 0; // Number of magical essences in inventory.
-		int invTools    = 0; // Number of crafting tools in inventory
-		int invVnum     = 0; // Vnum of the current inventory object.
-		for (P_obj inventory = ch->carrying; inventory; inventory = inventory->next_content)
-		{
-			invVnum = OBJ_VNUM(inventory);
-
-			if (invVnum == lowQualityMaterialVnum)
-				invLowMats++;
-			else if (invVnum == highQualityMaterialVnum)
-				invHighMats++;
-			else if (invVnum == VOBJ_CRAFTING_ESSENCE)
-				invEssences++;
-			else if (invVnum == VOBJ_CRAFTING_TOOLS)
-				invTools++;
-		}
-
-		// Check mats in inventory vs mats needed.
-		if (invLowMats < numLowest || invHighMats < numHighest)
-		{
-			send_to_char("You do not have the required &+ysalvaged &+Ymaterials &nin your inventory.\r\n", ch);
-			extract_obj(tobj);
-			extract_obj(matLowest);
-			extract_obj(matHighest);
-			return;
-		}
-		// If for some reason we want more than 1 box of tools, change the 1 below.
-		if (invTools < 1)
-		{
-			send_to_char("You must have &+ma &+ybox &+mof &+Rgnomish &+rcrafting &+mtools&n to create your item.\r\n", ch);
-			extract_obj(tobj);
-			extract_obj(matLowest);
-			extract_obj(matHighest);
-			return;
-		}
-		// If we're going to require multiple essences, need to edit this if statement.
-		if (invEssences < 1 && hasAffect)
-		{
-			send_to_char("You must have &+W1 &nof &+ma &+Mm&+Ya&+Mg&+Yi&+Mc&+Ya&+Ml &+messence&n due to the &+mmagical &nproperties this item possesses.\r\n", ch);
-			extract_obj(tobj);
-			extract_obj(matLowest);
-			extract_obj(matHighest);
-			return;
-		}
-
-		// Remove the materials from inventory...  Since we are changing the inventory
-		//   list, we need to use nextObj instead of just going to inventory->next_content.
-		P_obj nextObj;
-		// If for some reason we want more than 1 box of tools, change gotTools from boolean to int...
-		//   For right now, gotTools -> TRUE when we extract a set of tools from inventory.
-		bool gotTools = FALSE;
-		for (P_obj inventory = ch->carrying; inventory; inventory = nextObj)
-		{
-			nextObj = inventory->next_content;
-			invVnum = OBJ_VNUM(inventory);
-
-			if ((numLowest > 0) && (invVnum == lowQualityMaterialVnum))
-			{
-				obj_from_char(inventory);
-				extract_obj(inventory);
-				numLowest--;
-			}
-			else if ((numHighest > 0) && (invVnum == highQualityMaterialVnum))
-			{
-				obj_from_char(inventory);
-				extract_obj(inventory);
-				numHighest--;
-			}
-			// If we're requiring multiple essences, need to change this if clause.
-			else if (hasAffect && (invVnum == VOBJ_CRAFTING_ESSENCE))
-			{
-				obj_from_char(inventory);
-				extract_obj(inventory);
-				hasAffect = FALSE;
-			}
-			// If we're requiring multiple tools, need to change this if clause.
-			else if (!gotTools && (invVnum == VOBJ_CRAFTING_TOOLS))
-			{
-				obj_from_char(inventory);
-				extract_obj(inventory);
-				gotTools = TRUE;
-			}
-		}
-
-		notch_skill(ch, SKILL_CRAFT, 50);
-
-		SET_BIT(tobj->extra2_flags, ITEM2_CRAFTED);
-		SET_BIT(tobj->extra_flags, ITEM_NOREPAIR);
-		REMOVE_BIT(tobj->extra_flags, ITEM_SECRET);
-		randomizeitem(ch, tobj);
-		snprintf(keywords, MAX_STRING_LENGTH, "%s %s tradeskill", tobj->name, GET_NAME(ch));
-
-		snprintf(tempdesc, MAX_STRING_LENGTH, "%s", tobj->short_description);
-		snprintf(short_desc, MAX_STRING_LENGTH, "%s &+ymade by&n &+r%s&n", tempdesc, GET_NAME(ch));
-		set_keywords(tobj, keywords);
-		set_short_description(tobj, short_desc);
-
-		// Rewards here: tobj is the crafted item, so no need to load another.
-		wizlog(56, "%s crafted '%s' (%d) ival %d.", GET_NAME(ch), tobj->short_description, selected, itemvalue(tobj));
-
-		obj_to_char(tobj, ch);
-		act("&+W$n &+Ldelicately opens their &+ybox &+mof &+Rgnomish &+rcrafting &+mtools&+L and starts their work...\r\n"
-		    "&+W$n &+Lremoves the &+Wim&+wpur&+Lities &+Lfrom their &+ymaterials &+Land gently assembles a masterpiece...\r\n"
-		    "&+L...hands shaking, &+W$n &+Lraises their head and &+Ysmiles&+L, admiring their new $p.&N",
-		    TRUE,
-		    ch,
-		    tobj,
-		    0,
-		    TO_ROOM);
-		act("You &+Ldelicately open your &+ybox &+mof &+Rgnomish &+rcrafting &+mtools&+L and get to work...\r\n"
-		    "you &+Lremove the &+Wim&+wpur&+Lities &+Lfrom your &+ymaterials &+Land gently assemble a masterpiece...\r\n"
-		    "&+L...hands shaking, &+Wyou &+Lraise your head and &+Ysmile&+L, admiring your new $p.&N",
-		    FALSE,
-		    ch,
-		    tobj,
-		    0,
-		    TO_CHAR);
-
-		gain_exp(ch, NULL, iVal * 1000, EXP_BOON);
-		extract_obj(matLowest);
-		extract_obj(matHighest);
-		// Save the character! 1 -> in game.
-	if (!do_save_silent(ch, 1))
-		logit(LOG_DEBUG, "Failed to save %s after heroics reward.", GET_NAME(ch));
-	}
-
-	/*
-	   argument = one_argument(argument, Gbuf1);
-	   howmany = 0;
-	   i = 0;
-	   craft_obj1 = 0;
-	   craft_obj2 = 0;
-	   craft_obj3 = 0;
-
-	   for (t_obj = ch->carrying; t_obj; t_obj = nextobj)
-	   {
-	   nextobj = t_obj->next_content;
-
-
-	   if (isname("piece", t_obj->name) && !isname("tail", t_obj->name) &&
-	   obj_index[t_obj->R_num].virtual_number == RANDOM_EQ_VNUM)
-	   {
-	   i++;
-
-	   if (i == 1)
-	   {
-	   craft_obj1 = t_obj;
-	   }
-	   if (i == 2)
-	   {
-	   craft_obj2 = t_obj;
-	   }
-	   if (i == 3)
-	   {
-	   craft_obj3 = t_obj;
-	   }
-	   }
-
-	   }
-	   slot = -1;
-
-	   while (*argument == ' ')
-	   argument++;
-
-
-	   for (i = 0; i < MAX_SLOT; i++)
-	   {
-	   snprintf(Gbuf2, MAX_STRING_LENGTH, "%s\0", strip_ansi(slot_data[i].m_name).c_str());
-	   if (!strcmp(argument, Gbuf2))
-	   {
-	   howmany = slot_data[i].numb_material;
-	   slot = i;
-	   break;
-	   }
-	   }
-
-	   if (slot == -1 || howmany > 3)
-	   {
-	   act("You don't know how to create that...", FALSE, ch, 0, 0, TO_CHAR);
-	   return;
-	   }
-	   if (craft_obj1 == 0)
-	   {
-	   act("You do not have enough material, need to have a piece in inventory",
-	   FALSE, ch, 0, 0, TO_CHAR);
-	   return;
-	   }
-	   if (craft_obj2 == 0 && howmany == 2)
-	   {
-	   act
-	   ("You do not have enough material, need to have two pieces in inventory.",
-	   FALSE, ch, 0, 0, TO_CHAR);
-	   return;
-	   }
-	   if (craft_obj3 == 0 && howmany == 3)
-	   {
-	   act
-	("You do not have enough material need to have three pieces in inventory.",
-	 FALSE, ch, 0, 0, TO_CHAR);
-	return;
-  }
-
-
-  if (howmany == 2)
-  {
-	if (craft_obj1->material != craft_obj2->material)
-	{
-	  act("Mixing materials like that might be dangerous.", FALSE, ch, 0, 0,
-	      TO_CHAR);
-	  return;
-	}
-  }
-
-  if (howmany == 3)
-  {
-	if (craft_obj1->material != craft_obj2->material ||
-	    craft_obj1->material != craft_obj3->material)
-	{
-	  act("Mixing materials like that might be dangerous.", FALSE, ch, 0, 0,
-	      TO_CHAR);
-	  return;
-	}
-  }
-
-  for (i = 0; i <= MAXMATERIAL; i++)
-  {
-	if (material_data[i].m_number == craft_obj1->material)
-	{
-	  material_type = i;
-	}
-  }
-
-
-  */
-	/*        wizlog(56, "%s crafted.", GET_NAME(ch));
-
-	          if (howmany > 0)
-	          extract_obj(craft_obj1);
-	          if (howmany > 1)
-	          extract_obj(craft_obj2);
-	          if (howmany > 2)
-	          extract_obj(craft_obj3);
-	          obj = create_random_eq_new(ch, ch, slot, material_type);
-	          if (!obj)
-	          {
-	          act("&+yCrafting NOT COMPLETED CONTACT A GOD!&n", FALSE, ch, 0, 0,
-	          TO_CHAR);
-	          return;
-	          }
-	          obj_to_char(obj, ch);
-	//notch_skill(ch, SKILL_CRAFT, 30);
-	act("&+W$n crafted a $q&N.", TRUE, ch, obj, 0, TO_NOTVICT);
-	act("&+WYou crafted a $q&N", TRUE, ch, obj, 0, TO_CHAR);
-
-	return;
-	*/
+	crafting_handle_command(ch, CRAFTING_MODE_CRAFT, argument);
 }
+
 
 void do_smith(P_char ch, char *argument, int cmd) {}
