+++ src/crafting.h
@@ -22,6 +22,9 @@ struct crafting_plan
 /* Loads no player state. Returns FALSE when the item has no salvage family. */
 bool crafting_build_plan(P_obj item, struct crafting_plan *plan);
 int crafting_level_gate_multiplier(void);
+bool crafting_mode_enabled(enum crafting_mode mode);
+int crafting_essence_vnum(enum crafting_mode mode);
+int crafting_tool_vnum(enum crafting_mode mode);
 int *crafting_get_player_recipes(P_char ch, int *count);
 
 /* Reserved public module boundaries for the staged command/config extraction. */
