+++ src/tradeskill.c
@@ -377,354 +377,9 @@ P_obj check_foundry(P_char ch)
 
 void do_forge(P_char ch, char *argument, int cmd)
 {
-	int   skillLevel, objVnum, recipeNumber, commandType, iVal, invVnum;
-	int   numHighQuality, numLowQuality, lowQualityMaterialVnum, highQualityMaterialVnum;
-	int   invLowMats, invHighMats, invEssences;
-	char  buf[256];
-	char  recipe[256];
-	char  Gbuf1[MAX_STRING_LENGTH];
-	char  first[MAX_INPUT_LENGTH];
-	char  second[MAX_INPUT_LENGTH];
-	char  keywords[MAX_STRING_LENGTH];
-	char  short_desc[MAX_STRING_LENGTH];
-	char *rest;
-	bool  hasAffect, hasFlux;
-	P_obj obj, lowQualityMaterial, highQualityMaterial, inventory, invNextObj;
-
-	if (!(skillLevel = GET_CHAR_SKILL(ch, SKILL_FORGE)))
-	{
-		act("You do not know how to &+Lforge&n.", FALSE, ch, 0, 0, TO_CHAR);
-		return;
-	}
-
-	// load recipes from database
-	int  recipe_count = 0;
-	int *recipes      = sql_get_player_recipes(GET_PID(ch), &recipe_count);
-	if (!recipes || recipe_count == 0)
-	{
-		send_to_char("You dont know any recipes yet.\r\n", ch);
-		if (recipes)
-			free(recipes);
-		return;
-	}
-	rest = one_argument(argument, first);
-	one_argument(rest, second);
-	objVnum = atoi(second);
-
-	// If we don't have a first argument.
-	if (!*first)
-	{
-		commandType = 0;
-	}
-	else if (is_abbrev(first, "info"))
-	{
-		commandType = 1;
-	}
-	else if (is_abbrev(first, "stat"))
-	{
-		commandType = 2;
-	}
-	else if (is_abbrev(first, "make"))
-	{
-		commandType = 3;
-	}
-	// The first argument is invalid.
-	else
-	{
-		commandType = 0;
-	}
-
-	if (commandType == 0)
-	{
-		send_to_char("&+wForge Syntax:\n&n", ch);
-		send_to_char("&+w(forge info <number> - list required materials to forge the item.)\n&n", ch);
-		send_to_char("&+w(forge stat <number> - display properties of the item.)\n&n", ch);
-		send_to_char("&+w(forge make <number> - create the item.)\n&n", ch);
-		send_to_char("&+yYou know the following recipes:\n&n", ch);
-		send_to_char("----------------------------------------------------------------------------\n", ch);
-		send_to_char("&+BRecipe Number		              &+MItem&n\n\r", ch);
-
-		for (int i = 0; i < recipe_count; i++)
-		{
-			if (!(obj = read_object(recipes[i], VIRTUAL)))
-			{
-				logit(LOG_DEBUG, "'%s' has bad recipe vnum %d.", ch ? J_NAME(ch) : "NULL", recipes[i]);
-				continue;
-			}
-			snprintf(recipe, 256, "   &+W%-22d&n%s&n\n", recipes[i], obj->short_description);
-			page_string(ch->desc, recipe, 1);
-			send_to_char("----------------------------------------------------------------------------\n", ch);
-			extract_obj(obj);
-		}
-		free(recipes);
-		return;
-	}
-
-	// If the second argument doesn't exist or isn't a positive integer.
-	if (objVnum <= 0)
-	{
-		if (commandType == 1)
-		{
-			send_to_char("What &+Wrecipe&n would you like &+yinformation&n about?\n", ch);
-		}
-		else if (commandType == 2)
-		{
-			send_to_char("What &+Wrecipe&n would you like &+ystatistics&n about?\n", ch);
-		}
-		else
-		{
-			send_to_char("What &+Witem &nare you attempting to forge?\n", ch);
-		}
-		free(recipes);
-		return;
-	}
-
-	// Find objVnum in the recipe list
-	bool found = false;
-	for (int i = 0; i < recipe_count; i++)
-	{
-		if (recipes[i] == objVnum)
-		{
-			found = true;
-			break;
-		}
-	}
-	free(recipes);
-
-	// If we couldn't find the vnum in the recipe book (commandType is irrelevant).
-	if (!found)
-	{
-		send_to_char("You dont appear to have that &+Wrecipe&n in your list.&n\n", ch);
-		return;
-	}
-
-	// Attempt to load the object we're inspecting/making.
-	if (!(obj = read_object(objVnum, VIRTUAL)))
-	{
-		snprintf(Gbuf1, MAX_STRING_LENGTH, "Your recipe # %d seems to be &+rcorrupted&n. Please tell a &+WGod.\n\r", objVnum);
-		logit(LOG_DEBUG, "do_forge: '%s' has bad recipe vnum (%d) - couldn't load object.", ch ? J_NAME(ch) : "NULL", objVnum);
-		return;
-	}
-
-	// 1 -> "info"
-	if (commandType == 1)
-	{
-		// Display required materials to make obj - formula below:
-		struct crafting_plan plan;
-		if (!crafting_build_plan(obj, &plan))
-		{
-			send_to_char("You couldn't figure out what materials to use.\n\r", ch);
-			extract_obj(obj);
-			return;
-		}
-		iVal                    = plan.item_value;
-		numHighQuality          = plan.high_material_count;
-		numLowQuality           = plan.low_material_count;
-		lowQualityMaterialVnum  = plan.low_material_vnum;
-		highQualityMaterialVnum = plan.high_material_vnum;
-		lowQualityMaterial      = read_object(lowQualityMaterialVnum, VIRTUAL);
-		highQualityMaterial     = read_object(highQualityMaterialVnum, VIRTUAL);
-
-		if (lowQualityMaterial == NULL || highQualityMaterial == NULL)
-		{
-			send_to_char("You couldn't figure out what materials to use.\n\r", ch);
-			if (lowQualityMaterial != NULL)
-			{
-				extract_obj(lowQualityMaterial);
-			}
-			if (highQualityMaterial != NULL)
-			{
-				extract_obj(highQualityMaterial);
-			}
-			extract_obj(obj);
-			return;
-		}
-
-		send_to_char("&+yYou open your &+Ltome &+yof &+Ycra&+yftsm&+Lanship &+yand examine the &+Litem&n.\n", ch);
-		if (numLowQuality == 0)
-		{
-			snprintf(recipe, MAX_STRING_LENGTH, "To forge this item, you will need %d of %s.\r\n&n", numHighQuality, highQualityMaterial->short_description);
-		}
-		else
-		{
-			// numHighQuality will always be >= 1 with the code the way it is on 2/11/2015.
-			if (numHighQuality == 0)
-			{
-				snprintf(recipe, MAX_STRING_LENGTH, "To forge this item, you will need %d of %s.\r\n&n", numLowQuality, lowQualityMaterial->short_description);
-			}
-			else
-			{
-				snprintf(recipe,
-				         MAX_STRING_LENGTH,
-				         "To forge this item, you will need %d of %s and %d of %s.\r\n&n",
-				         numHighQuality,
-				         highQualityMaterial->short_description,
-				         numLowQuality,
-				         lowQualityMaterial->short_description);
-			}
-		}
-		// If we're going to require multiple essences, need to edit this if statement.
-		if (has_affect(obj))
-		{
-			strcat(recipe, "You must have &+W1 &nof &+ma &+Mm&+Ya&+Mg&+Yi&+Mc&+Ya&+Ml &+messence&n due to the &+mmagical &nproperties this item possesses.\r\n");
-		}
-
-		page_string(ch->desc, recipe, 1);
-		extract_obj(obj);
-		extract_obj(highQualityMaterial);
-		extract_obj(lowQualityMaterial);
-		return;
-	}
-	// 2 -> "stat"
-	else if (commandType == 2)
-	{
-		// Display info on obj:
-		send_to_char("&+yYou open your &+Ltome &+yof &+Ycra&+yftsm&+Lanship &+yand examine the &+Litem&n.\n", ch);
-		spell_identify(GET_LEVEL(ch), ch, 0, 0, 0, obj);
-		extract_obj(obj);
-		return;
-	}
-	// 3 -> "make"
-	else if (commandType == 3)
-	{
-		// Attempt to make obj:
-		struct crafting_plan plan;
-		if (!crafting_build_plan(obj, &plan))
-		{
-			send_to_char("You couldn't figure out what materials to use.\n\r", ch);
-			extract_obj(obj);
-			return;
-		}
-		iVal = plan.item_value;
-		if (iVal > GET_LEVEL(ch) * crafting_level_gate_multiplier() || IS_OBJ_STAT2(obj, ITEM2_QUESTITEM))
-		{
-			act("You look at the recipe for $p&n, but can't seem to discern how to make it.  &+mHow strange.&N", FALSE, ch, obj, 0, TO_CHAR);
-			extract_obj(obj);
-			return;
-		}
-		numHighQuality          = plan.high_material_count;
-		numLowQuality           = plan.low_material_count;
-		lowQualityMaterialVnum  = plan.low_material_vnum;
-		highQualityMaterialVnum = plan.high_material_vnum;
-		hasAffect               = plan.magical;
-
-		/* Foundry code here (not requiring one atm I guess):
-		if( !check_foundry(ch) )
-		{
-		  act("&+LYou need to be by your foundry to forge...&n", FALSE, ch, 0, 0, TO_CHAR);
-		  extract_obj( obj );
-		  return;
-		}
-		*/
-
-		hasFlux    = FALSE;
-		invLowMats = invHighMats = invEssences = 0;
-		// Count up the materials ch has on hand.
-		for (inventory = ch->carrying; inventory; inventory = inventory->next_content)
-		{
-			invVnum = OBJ_VNUM(inventory);
-
-			if (invVnum == lowQualityMaterialVnum)
-			{
-				invLowMats++;
-			}
-			else if (invVnum == highQualityMaterialVnum)
-			{
-				invHighMats++;
-			}
-			else if (invVnum == VOBJ_FORGING_ESSENCE)
-			{
-				invEssences++;
-			}
-			else if (invVnum == VOBJ_FORGING_FLUX)
-			{
-				hasFlux = TRUE;
-			}
-		}
-
-		// Check to make sure ch has enough materials.
-		if (hasFlux == FALSE)
-		{
-			send_to_char("You must have a &+Lblacksmithing &nflux to complete the smithing process!\r\n", ch);
-			extract_obj(obj);
-			return;
-		}
-		if ((hasAffect && invEssences < 1) || (invLowMats < numLowQuality) || (invHighMats < numHighQuality))
-		{
-			send_to_char("You do not have the required &+ysalvaged &+Ymaterials &nin your inventory.\r\n", ch);
-			extract_obj(obj);
-			return;
-		}
-
-		// Ok, ch has the materials needed to create obj in inventory.. Now take them away, muahahah!
-		for (inventory = ch->carrying; inventory; inventory = invNextObj)
-		{
-			invNextObj = inventory->next_content;
-			invVnum    = OBJ_VNUM(inventory);
-
-			if ((invVnum == lowQualityMaterialVnum) && (numLowQuality > 0))
-			{
-				extract_obj(inventory, TRUE); // Not an arti, but 'in game.'
-				numLowQuality--;
-			}
-			else if ((invVnum == highQualityMaterialVnum) && (numHighQuality > 0))
-			{
-				extract_obj(inventory, TRUE); // Not an arti, but 'in game.'
-				numHighQuality--;
-			}
-			else if (hasAffect && (invVnum == VOBJ_FORGING_ESSENCE))
-			{
-				extract_obj(inventory, TRUE); // Not an arti, but 'in game.'
-				hasAffect = FALSE;
-			}
-			else if (hasFlux && (invVnum == VOBJ_FORGING_FLUX))
-			{
-				extract_obj(inventory, TRUE); // Not an arti, but 'in game.'
-				hasFlux = FALSE;
-			}
-		}
-
-		notch_skill(ch, SKILL_FORGE, 50);
-		SET_BIT(obj->extra2_flags, ITEM2_CRAFTED);
-		SET_BIT(obj->extra_flags, ITEM_NOREPAIR);
-		REMOVE_BIT(obj->extra_flags, ITEM_SECRET);
-		randomizeitem(ch, obj);
-
-		snprintf(keywords, MAX_STRING_LENGTH, "%s %s tradeskill", obj->name, GET_NAME(ch));
-		snprintf(short_desc, MAX_STRING_LENGTH, "%s &+ymade by&n &+r%s&n", obj->short_description, GET_NAME(ch));
-		set_keywords(obj, keywords);
-		set_short_description(obj, short_desc);
-
-		wizlog(56, "%s forged '%s' (%d) ival %d.", GET_NAME(ch), obj->short_description, objVnum, itemvalue(obj));
-		obj_to_char(obj, ch);
-
-		act("&+W$n &+Lgently takes their &+ymaterials&+L, their &nflux&+L, and places them into the &+rf&+Ro&+Yr&+Rg&+re&+L.\r\n"
-		    "&+W$n &+Lremoves the &+yitems &+Lfrom the &+rheat &+Land starts to &nhammer &+Laway at the mixture..\r\n"
-		    "&+L...after shedding plenty of &+Wsweat&+L, &+W$n &+Lsteps back, admiring their new $p.&N",
-		    TRUE,
-		    ch,
-		    obj,
-		    0,
-		    TO_ROOM);
-		act("You &+Lgently take your &+ymaterials&+L, the &nflux&+L, and place them into the &+rf&+Ro&+Yr&+Rg&+re&+L.\r\n"
-		    "You &+Lremove the &+yitems &+Lfrom the &+rheat &+Land start to &nhammer &+Laway at the mixture..\r\n"
-		    "&+L...after shedding plenty of &+Wsweat&+L, you &+Lstep back, admiring your new $p.&N",
-		    FALSE,
-		    ch,
-		    obj,
-		    0,
-		    TO_CHAR);
-
-		gain_exp(ch, NULL, (iVal * 1000), EXP_BOON);
-	}
-	else
-	{
-		send_to_char("&+RBad command type&n: Please report this to an Immortal.\n\r", ch);
-		logit(LOG_DEBUG, "do_forge: bad command arguments '%s' by '%s'.", argument, ch ? J_NAME(ch) : "NULL");
-		extract_obj(obj);
-		return;
-	}
+	crafting_handle_command(ch, CRAFTING_MODE_FORGE, argument);
 }
+
 
 /* OLD FORGE CODE:
   if (GET_CHAR_SKILL(ch, SKILL_FORGE) == 0){
