+++ tests/async/test_crafting_module_contract.py
@@ -1,17 +1,23 @@
 #!/usr/bin/env python3
-"""Modern Craft and Forge must share a config-backed execution owner."""
+"""Modern Craft and Forge must be dispatched by the dedicated module."""
 from pathlib import Path
 
 root = Path(__file__).resolve().parents[2]
-crafting_h = root / "src/crafting.h"
-crafting_c = root / "src/crafting.c"
-assert crafting_h.exists(), "missing src/crafting.h"
-assert crafting_c.exists(), "missing src/crafting.c"
-header = crafting_h.read_text()
-source = crafting_c.read_text()
-assert "void boot_crafting_system(void);" in header
-assert "crafting_handle_command" in header
+header = (root / "src/crafting.h").read_text()
+source = (root / "src/crafting.c").read_text()
+craft = (root / "src/actnew.c").read_text()
+forge = (root / "src/tradeskill.c").read_text()
+config = (root / "lib/crafting.cfg").read_text()
+
 assert "crafting_build_plan" in source
 assert "CRAFTING_MODE_CRAFT" in header
 assert "CRAFTING_MODE_FORGE" in header
+assert "crafting_handle_craft_command" in source
+assert "crafting_handle_forge_command" in source
+assert "crafting_handle_command(ch, CRAFTING_MODE_CRAFT, argument);" in craft
+assert "crafting_handle_command(ch, CRAFTING_MODE_FORGE, argument);" in forge
+assert "crafting.craft.enabled=1" in config
+assert "crafting.forge.enabled=1" in config
+assert "crafting_essence_vnum(CRAFTING_MODE_CRAFT)" in source
+assert "crafting_tool_vnum(CRAFTING_MODE_FORGE)" in source
 print("crafting module contract passed")
